Portfolio Exercise: Starbucks (5.5) WIP
Experiments with different scaling (Standard vs MinMax), oversampling (Random vs SMOTE) and models (added KNN).

Best model so far is...

LogisticRegression(random_state=42)
Your irr with this strategy is 0.0192.
Your nir with this strategy is 66.35.
We came up with a model with an irr of 0.0188 and an nir of 189.45 on the test set.